// src/app/OpenAPI/model/characteristicsTypeDTO.ts
/**
 * SportBuddy
 * API specification for my thesis work SportBuddy
 *
 * The version of the OpenAPI document: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CharacteristicsTypeDTO = 'arms' | 'legs';

export const CharacteristicsTypeDTO = {
    Arms: 'arms' as CharacteristicsTypeDTO,
    Legs: 'legs' as CharacteristicsTypeDTO
};