// src/app/OpenAPI/model/distanceTypeDTO.ts
/**
 * SportBuddy
 * API specification for my thesis work
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: abc@def.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type DistanceTypeDTO = 'around0' | 'around1' | 'around2' | 'around3' | 'around4' | 'aroundMinus1' | 'aroundMinus2' | 'aroundMinus3' | 'aroundMinus4';

export const DistanceTypeDTO = {
    Around0: 'around0' as DistanceTypeDTO,
    Around1: 'around1' as DistanceTypeDTO,
    Around2: 'around2' as DistanceTypeDTO,
    Around3: 'around3' as DistanceTypeDTO,
    Around4: 'around4' as DistanceTypeDTO,
    AroundMinus1: 'aroundMinus1' as DistanceTypeDTO,
    AroundMinus2: 'aroundMinus2' as DistanceTypeDTO,
    AroundMinus3: 'aroundMinus3' as DistanceTypeDTO,
    AroundMinus4: 'aroundMinus4' as DistanceTypeDTO
};