// src/app/OpenAPI/model/fullPositionTypeDTO.ts
/**
 * SportBuddy
 * API specification for my thesis work SportBuddy
 *
 * The version of the OpenAPI document: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type FullPositionTypeDTO = 'around0' | 'around45' | 'around90' | 'around135' | 'around180' | 'around225' | 'around270' | 'around315';

export const FullPositionTypeDTO = {
    Around0: 'around0' as FullPositionTypeDTO,
    Around45: 'around45' as FullPositionTypeDTO,
    Around90: 'around90' as FullPositionTypeDTO,
    Around135: 'around135' as FullPositionTypeDTO,
    Around180: 'around180' as FullPositionTypeDTO,
    Around225: 'around225' as FullPositionTypeDTO,
    Around270: 'around270' as FullPositionTypeDTO,
    Around315: 'around315' as FullPositionTypeDTO
};