// src/app/OpenAPI/model/halfPositionTypeDTO.ts
/**
 * SportBuddy
 * API specification for my thesis work SportBuddy
 *
 * The version of the OpenAPI document: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type HalfPositionTypeDTO = 'around0' | 'around45' | 'around90' | 'around135' | 'around180';

export const HalfPositionTypeDTO = {
    Around0: 'around0' as HalfPositionTypeDTO,
    Around45: 'around45' as HalfPositionTypeDTO,
    Around90: 'around90' as HalfPositionTypeDTO,
    Around135: 'around135' as HalfPositionTypeDTO,
    Around180: 'around180' as HalfPositionTypeDTO
};